backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
passlib[bcrypt]
python-jose[cryptography]
